+++ src/Http/Models/Group.php
@@ -5,10 +5,12 @@ namespace Junges\ACL\Http\Models;
 use Junges\ACL\Events\GroupSaving;
 use Junges\ACL\Traits\GroupsTrait;
 use Illuminate\Database\Eloquent\Model;
+use Junges\ACL\Traits\ACLWildcardsTrait;
 
 class Group extends Model
 {
     use GroupsTrait;
+    use ACLWildcardsTrait;
 
     protected $dates = ['deleted_at'];
     protected $table;