+++ tests/User.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Junges\ACL\Test;
+namespace Junges\ACL\Tests;
 
 use Junges\ACL\Traits\UsersTrait;
 use Illuminate\Notifications\Notifiable;
+use Junges\ACL\Traits\ACLWildcardsTrait;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 
 class User extends Authenticatable
 {
     use UsersTrait;
+    use ACLWildcardsTrait;
     use Notifiable;
     /**
      * The attributes that are mass assignable.